requirements.txt
Flask==3.1.0
gunicorn==23.0
pandas==2.2.3
scikit-learn==1.6.1
python-dotenv==1.0.1
pytest==8.3.5